pyproject.toml
[tool.poetry]
name = "django-todo"
version = "0.1.0"
description = "Django To Do app"
authors = ["Dan Sto. Domingo <danielesto.domingo@gmail.com>"]
readme = "README.md"
packages = [{include = "django_todo"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
django-rest-framework = "^0.1.0"
drf-spectacular = "^0.25.1"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.5"
djangorestframework-camel-case = "^1.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
